refactoring of RandomJokeCrawler: splitting app into core and console module. Implementing a repeating dialogue for new jokes. Javadoc comments of methods.
