+++ games/pom.xml
@@ -28,18 +28,17 @@
     <modules>
         <module>gameoflife-console</module>
         <module>gameoflife-core</module>
+        <module>gameoflife-gui</module>
         <module>guessnumber-console</module>
         <module>guessnumber-core</module>
-        <module>tictactoe-console</module>
-        <module>tictactoe-core</module>
+        <module>randomjokecrawler-console</module>
+        <module>randomjokecrawler-core</module>
         <module>start-console</module>
         <module>start-core</module>
         <module>start-gui</module>
-        <module>gameoflife-gui</module>
+        <module>tictactoe-core</module>
+        <module>tictactoe-console</module>
         <module>tictactoe-gui</module>
-        <module>random-joke-crawler</module>
-        <module>randomjokecrawler-core</module>
-        <module>randomejokecrawler-console</module>
     </modules>
