+++ games/start-console/pom.xml
@@ -26,13 +26,6 @@
         <!-- Transitive Dependency -->
         <dependency>
             <groupId>ch.zhaw</groupId>
-            <artifactId>start-core</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <!-- Transitive Dependency -->
-        <dependency>
-            <groupId>ch.zhaw</groupId>
             <artifactId>gameoflife-console</artifactId>
             <version>${project.version}</version>
         </dependency>
@@ -47,14 +40,21 @@
         <!-- Transitive Dependency -->
         <dependency>
             <groupId>ch.zhaw</groupId>
-            <artifactId>tictactoe-console</artifactId>
+            <artifactId>randomejokecrawler-console</artifactId>
             <version>${project.version}</version>
         </dependency>
 
         <!-- Transitive Dependency -->
         <dependency>
             <groupId>ch.zhaw</groupId>
-            <artifactId>randomejokecrawler-console</artifactId>
+            <artifactId>start-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <!-- Transitive Dependency -->
+        <dependency>
+            <groupId>ch.zhaw</groupId>
+            <artifactId>tictactoe-console</artifactId>
             <version>${project.version}</version>
         </dependency>
